PSQLADM-131 : Fixup scripts for 8.0
Issue
Scripts are not woring with PXC 8.0

Solution
Many fixes. Mostly due to authentication changes with 8.0
For instance, the password (authentication_string) field can now contain
binary strings (so unprintable characters), so there is a need
to properly escape the strings and to properly iterate over the
return values from queries.